+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/flat-view.type.ts
@@ -12,13 +12,3 @@ export type FlatView = Pick<
 > & {
   viewFields: FlatViewField[];
 };
-
-export type CreateViewAction = {
-  type: 'create_view';
-  view: FlatView;
-};
-
-export type WorkspaceMigrationViewActionV2 = CreateViewAction;
-
-export type WorkspaceMigrationViewActionTypeV2 =
-  WorkspaceMigrationViewActionV2['type'];
