+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/__tests__/common/workspace-migration-builder-object-test-case.ts
@@ -81,6 +81,7 @@ const CREATE_OBJECT_TEST_CASES: WorkspaceMigrationBuilderTestCase[] = [
       },
       expectedActionsTypeCounter: {
         createObject: 1,
+        createView: 1,
       },
     },
   },
@@ -102,6 +103,7 @@ const CREATE_OBJECT_TEST_CASES: WorkspaceMigrationBuilderTestCase[] = [
       },
       expectedActionsTypeCounter: {
         createObject: 1,
+        createView: 1,
       },
     },
   },
@@ -112,6 +114,7 @@ const CREATE_OBJECT_TEST_CASES: WorkspaceMigrationBuilderTestCase[] = [
       expectedActionsTypeCounter: {
         createIndex: 1,
         createObject: 1,
+        createView: 1,
       },
       input: {
         fromFlatObjectMetadataMaps:
