+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/__tests__/workspace-migration-builder.spec.ts
@@ -54,6 +54,7 @@ const expectedActionsTypeCounterChecker = ({
     deleteObject: 0,
     updateField: 0,
     updateObject: 0,
+    createView: 0,
   };
   const actualActionsTypeCounter = workspaceMigration.actions.reduce(
     (acc, action) => {
