+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-view-action-v2.type.ts
@@ -1,0 +1,11 @@
+import { type FlatView } from 'src/engine/workspace-manager/workspace-migration-v2/types/flat-view.type';
+
+export type CreateViewAction = {
+  type: 'create_view';
+  view: FlatView;
+};
+
+export type WorkspaceMigrationViewActionV2 = CreateViewAction;
+
+export type WorkspaceMigrationViewActionTypeV2 =
+  WorkspaceMigrationViewActionV2['type'];
